# [Enhancement] Support multi partition columns in global runtime filter