--- a/be/src/exec/pipeline/operator.cpp
+++ b/be/src/exec/pipeline/operator.cpp
@@ -186,6 +186,8 @@ void Operator::_init_rf_counters(bool init_bloom) {
     }
     if (init_bloom && _bloom_filter_eval_context.join_runtime_filter_timer == nullptr) {
         _bloom_filter_eval_context.join_runtime_filter_timer = ADD_TIMER(_common_metrics, "JoinRuntimeFilterTime");
+        _bloom_filter_eval_context.join_runtime_filter_hash_timer =
+                ADD_TIMER(_common_metrics, "JoinRuntimeFilterHashTime");
         _bloom_filter_eval_context.join_runtime_filter_input_counter =
                 ADD_COUNTER(_common_metrics, "JoinRuntimeFilterInputRows", TUnit::UNIT);
         _bloom_filter_eval_context.join_runtime_filter_output_counter =
@@ -229,6 +231,7 @@ Status OperatorFactory::prepare(RuntimeState* state) {
             if (grf == nullptr) {
                 continue;
             }
+
             desc->set_shared_runtime_filter(grf);
         }
     }