--- a/be/src/exec/vectorized/project_node.cpp
+++ b/be/src/exec/vectorized/project_node.cpp
@@ -263,6 +263,9 @@ void ProjectNode::push_down_join_runtime_filter(RuntimeState* state,
     auto iter = descriptors.begin();
     while (iter != descriptors.end()) {
         RuntimeFilterProbeDescriptor* rf_desc = iter->second;
+        if (!rf_desc->can_push_down_runtime_filter()) {
+            continue;
+        }
         SlotId slot_id;
         // bound to this tuple and probe expr is slot ref.
         if (!rf_desc->is_bound(_tuple_ids) || !rf_desc->is_probe_slot_ref(&slot_id)) {