@wanpengfei-git 2022-08-19
#### [FE PR Coverage Check]

:disappointed: **fail** : 11 / 29 (37.93%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/RuntimeFilterDescription.java|1|5|20.00%|[228, 232, 233, 289]|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|7|18|38.89%|[740, 741, 744, 745, 749, 750, 751, 753, 754, 756, 764]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|3|6|50.00%|[1079, 1080, 1083]|