@wanpengfei-git 2022-08-20
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 26 / 32 (81.25%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/RuntimeFilterDescription.java|3|5|60.00%|[228, 289]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|4|6|66.67%|[1079, 1080]|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|19|21|90.48%|[276, 282]|