@wanpengfei-git 2022-08-23
#### [FE PR Coverage Check]

:disappointed: **fail** : 23 / 34 (67.65%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|4|11|36.36%|[738, 739, 743, 744, 745, 747, 749]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|4|6|66.67%|[1092, 1093]|
|:large_blue_circle:|com/starrocks/planner/RuntimeFilterDescription.java|4|5|80.00%|[228]|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|11|12|91.67%|[255]|