@LiShuMing 2022-08-24
run starrocks_admit_test