@wanpengfei-git 2022-08-24
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 24 / 30 (80.00%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/RuntimeFilterDescription.java|3|5|60.00%|[228, 289]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|4|6|66.67%|[1092, 1093]|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|14|16|87.50%|[748, 754]|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|3|3|100.00%|[]|