@wanpengfei-git 2022-08-24
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 31 / 37 (83.78%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/RuntimeFilterDescription.java|3|5|60.00%|[228, 289]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|8|10|80.00%|[1100, 1101]|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|14|16|87.50%|[748, 754]|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|6|6|100.00%|[]|