@wangruin 2022-08-27
run starrocks_fe_unittest