@wanpengfei-git 2022-08-29
#### [FE PR Coverage Check]

:disappointed: **fail** : 192 / 244 (78.69%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/JoinNode.java|26|40|65.00%|[216, 229, 232, 259, 260, 261, 262, 263, 265, 266, 267, 276, 290, 312]|
|:large_blue_circle:|com/starrocks/planner/RuntimeFilterDescription.java|6|9|66.67%|[133, 291, 292]|
|:large_blue_circle:|com/starrocks/planner/SetOperationNode.java|28|41|68.29%|[245, 248, 259, 280, 281, 282, 283, 284, 286, 287, 289, 303, 313]|
|:large_blue_circle:|com/starrocks/planner/AnalyticEvalNode.java|14|18|77.78%|[259, 262, 282, 288]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|8|10|80.00%|[1100, 1101]|
|:large_blue_circle:|com/starrocks/planner/AggregationNode.java|17|20|85.00%|[249, 256, 278]|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|70|81|86.42%|[743, 746, 758, 767, 821, 838, 851, 852, 853, 854, 855]|
|:large_blue_circle:|com/starrocks/planner/ProjectNode.java|15|17|88.24%|[132, 162]|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|8|8|100.00%|[]|