@LiShuMing 2022-08-29
run starrocks_admit_test