@LiShuMing 2022-08-30
run starrocks_fe_unittest