@kangkaisen 2022-08-30
> In tpcds-100g, this feature can optimize queries below:

What's the  tpcds-1T and   tpcds-10T behavior?